Introduction of BER types (ber_int_t, ber_tag_t, ber_len_t, ber_socket_t) and lber_types.h (NT version only, need to write configure support for other platforms).   Changes to sources are likely incomplete and buggy. This is a work in progress.

+++ servers/slapd/monitor.c
@@ -18,7 +18,7 @@
 #include <ac/string.h>
 #include <ac/time.h>
 
-#include "ldap_defaults.h"
+#include "ldapconfig.h"
 #include "slap.h"
 
 #if defined( SLAPD_MONITOR_DN )
@@ -120,7 +120,7 @@ monitor_info( Connection *conn, Operation *op )
 	val.bv_len = strlen( buf );
 	attr_merge( e, "totalconnections", vals );
 
-	sprintf( buf, "%d", dtblsize );
+	sprintf( buf, "%ld", (long) dtblsize );
 	val.bv_val = buf;
 	val.bv_len = strlen( buf );
 	attr_merge( e, "dtablesize", vals );
